requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
requests==2.32.3
python-dotenv==1.1.0
Flask-Login==0.6.3
pytest==8.3.4
